backend/pyproject.toml
[project]
name = "document-intelligence-backend"
version = "2.0.0"
description = "Kendi Dokümanların ile Sohbet Et — Google Gemini destekli RAG backend"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=0.6.3",
    "fastapi[standard]>=0.115.12",
    "langchain>=0.3.25",
    "langchain-chroma>=0.2.3",
    "langchain-google-genai>=2.1.2",
    "langgraph>=0.5.0",
    "mammoth>=1.9.0",
    "pydantic-settings>=2.9.1",
    "pypdf>=5.6.0",
    "python-docx>=1.1.2",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "uvicorn[standard]>=0.34.2",
]
